+++ src/main/java/com/wakeUpTogetUp/togetUp/api/mission/MissionService.java
@@ -10,9 +10,9 @@ import com.wakeUpTogetUp.togetUp.api.mission.model.CustomDetectedObject;
 import com.wakeUpTogetUp.togetUp.api.mission.model.CustomDetectedTag;
 import com.wakeUpTogetUp.togetUp.api.mission.model.Emotion;
 import com.wakeUpTogetUp.togetUp.api.mission.model.MissionLog;
-import com.wakeUpTogetUp.togetUp.api.mission.service.AzureVision32Service;
-import com.wakeUpTogetUp.togetUp.api.mission.service.AzureVision40Service;
-import com.wakeUpTogetUp.togetUp.api.mission.service.GoogleVisionService;
+import com.wakeUpTogetUp.togetUp.infra.azure.vision.AzureVision32Service;
+import com.wakeUpTogetUp.togetUp.infra.azure.vision.AzureVision40Service;
+import com.wakeUpTogetUp.togetUp.infra.google.vision.GoogleVisionService;
 import com.wakeUpTogetUp.togetUp.api.mission.service.ObjectDetectionService;
 import com.wakeUpTogetUp.togetUp.api.notification.NotificationService;
 import com.wakeUpTogetUp.togetUp.api.users.UserAvatarService;
