+++ src/main/java/com/wakeUpTogetUp/togetUp/api/mission/service/MissionImageService.java
@@ -3,7 +3,7 @@ package com.wakeUpTogetUp.togetUp.api.mission.service;
 import com.google.cloud.vision.v1.FaceAnnotation;
 import com.wakeUpTogetUp.togetUp.api.mission.model.CustomDetectedObject;
 import com.wakeUpTogetUp.togetUp.utils.ImageProcessor;
-import com.wakeUpTogetUp.togetUp.api.file.model.CustomFile;
+import com.wakeUpTogetUp.togetUp.infra.aws.s3.model.CustomFile;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.imaging.formats.tiff.TiffImageMetadata;
