+++ src/main/java/com/wakeUpTogetUp/togetUp/infra/aws/s3/FileService.java
@@ -1,11 +1,11 @@
-package com.wakeUpTogetUp.togetUp.api.file;
+package com.wakeUpTogetUp.togetUp.infra.aws.s3;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.wakeUpTogetUp.togetUp.infra.aws.s3.model.CustomFile;
 import com.wakeUpTogetUp.togetUp.utils.FileUtil;
-import com.wakeUpTogetUp.togetUp.api.file.model.CustomFile;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
